+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -41,6 +41,7 @@ import is.codion.swing.framework.model.SwingEntityTableModel;
 import is.codion.swing.framework.ui.icon.FrameworkIcons;
 
 import javax.swing.AbstractAction;
+import javax.swing.Action;
 import javax.swing.ImageIcon;
 import javax.swing.JComponent;
 import javax.swing.JPanel;
@@ -889,7 +890,10 @@ public class EntityPanel extends JPanel {
 	 */
 	protected final void initializeTablePanel() {
 		if (tablePanel != null) {
-			tablePanel.table().doubleClickAction().mapNull(() -> Control.control(new ShowHiddenEditPanel()));
+			Value<Action> doubleClickAction = tablePanel.table().doubleClickAction();
+			if (doubleClickAction.isNull()) {
+				doubleClickAction.set(Control.control(new ShowHiddenEditPanel()));
+			}
 			tablePanel.initialize();
 		}
 	}